Improves indentation to ignore extra spaces
When indenting code sometimes spaces are used besides tabs. This
improvement makes ignore those extra spaces to taken into account just
the tabs for the indentation of each line.

This also works when only spaces are used to indent because the current
indentation of a line is adjusted to a multiple of 'shiftwidth' to
ignore extra spaces.